requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
joblib==1.0.1
jsonpickle==2.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.19.5
Pillow==8.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
SQLAlchemy==1.4.13
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
urllib3==1.26.4
view==0.1
virtualenv==20.4.2
waitress==2.0.0
Werkzeug==1.0.1
wrapt==1.12.1
pypotrace==0.3
opencv-python==4.5.1.48
